requirements.txt
alabaster==0.7.16
antlr4-python3-runtime==4.11.1
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.20.1
git-me-the-url==2.1.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
peakrdl==1.1.0
peakrdl-cheader==1.0.0
-e git+ssh://git@github.com/Risto97/PeakRDL-halcpp.git@f656230a801142801c6ffd5eea93551d1373ec4f#egg=peakrdl_halcpp
peakrdl-html==2.10.1
peakrdl-ipxact==3.4.3
peakrdl-regblock==0.20.0
peakrdl-systemrdl==0.3.0
peakrdl-uvm==2.3.0
Pygments==2.17.2
python-markdown-math==0.8
requests==2.31.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
systemrdl-compiler==1.27.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
zipp==3.17.0
